+++ dockerfiles/mysql/bitbank.sql
@@ -18,6 +18,10 @@ CREATE TABLE `bank_account_transactions` (
 
 INSERT INTO `bank_account_transactions` (`id`, `amount`, `purpose`, `transaction_date`) VALUES
 (1,	2500.00,	'January 2022 salary',	'2022-02-06 16:49:03.000000'),
-(2,	-450.00,	'Arthur High School February 2022',	'2022-02-07 07:15:29.000000');
+(2,	-450.00,	'Arthur High School February 2022',	'2022-02-07 07:15:29.000000'),
+(3,	100.00,	'Postman test',	'2022-02-08 11:49:59.245823'),
+(4,	74.50,	'Postman test 2',	'2022-02-08 14:02:49.854420'),
+(5,	7.40,	'Daily Interest - CED 002',	'2022-02-10 07:49:32.120245'),
+(6,	4.40,	'Daily Interest - CED 002',	'2022-02-10 07:53:57.450064');
 
--- 2022-02-07 07:15:57
+-- 2022-02-10 10:39:25