+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Setup tests
++ Add CI/CD
 
 ## 0.1.0
 + First commit