+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Setup tests
 + Add CI/CD
++ Add Bank Account Transactions entity
 
 ## 0.1.0
 + First commit