+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Setup tests
 + Add CI/CD
 + Add Bank Account Transactions entity
++ Add Bank Account Transactions repository
 
 ## 0.1.0
 + First commit