+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Add Bank Account Transactions entity
 + Add Bank Account Transactions repository
 + Add Bank Account Transactions service
++ Add Bank Account Transactions controller + DTO
 
 ## 0.1.0
 + First commit