+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Add Bank Account Transactions repository
 + Add Bank Account Transactions service
 + Add Bank Account Transactions controller + DTO
++ Add Docker for deploy
 
 ## 0.1.0
 + First commit