+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 + Add Bank Account Transactions service POST
 + Add Bank Account Transactions controller POST
 + Add Ping service GET
++ Add Ping controller GET
 
 ## 0.1.0
 + First commit