+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 0.2.0
 + Setup tests
 + Add CI/CD
 + Add Bank Account Transactions entity
@@ -10,6 +10,7 @@
 + Add Bank Account Transactions controller POST
 + Add Ping service GET
 + Add Ping controller GET
++ Add exclusion in pom.xml for 100% coverage
 
 ## 0.1.0
 + First commit